chore: expose session id for react native
