+++ posthog/src/main/java/com/posthog/PostHog.kt
@@ -17,9 +17,9 @@ import com.posthog.internal.PostHogPrintLogger
 import com.posthog.internal.PostHogQueue
 import com.posthog.internal.PostHogSendCachedEventsIntegration
 import com.posthog.internal.PostHogSerializer
+import com.posthog.internal.PostHogSessionManager
 import com.posthog.internal.PostHogThreadFactory
 import com.posthog.vendor.uuid.TimeBasedEpochGenerator
-import java.util.UUID
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
 
@@ -48,13 +48,8 @@ public class PostHog private constructor(
     private val setupLock = Any()
     private val optOutLock = Any()
     private val anonymousLock = Any()
-    private val sessionLock = Any()
     private val groupsLock = Any()
 
-    // do not move to companion object, otherwise sessionId will be null
-    private val sessionIdNone = UUID(0, 0)
-
-    private var sessionId = sessionIdNone
     private val featureFlagsCalledLock = Any()
 
     private var config: PostHogConfig? = null
@@ -272,15 +267,13 @@ public class PostHog private constructor(
             }
         }
 
-        synchronized(sessionLock) {
-            if (sessionId != sessionIdNone) {
-                val sessionId = sessionId.toString()
-                props["\$session_id"] = sessionId
-                if (config?.sessionReplay == true) {
-                    // Session replay requires $window_id, so we set as the same as $session_id.
-                    // the backend might fallback to $session_id if $window_id is not present next.
-                    props["\$window_id"] = sessionId
-                }
+        PostHogSessionManager.getActiveSessionId()?.let { sessionId ->
+            val tempSessionId = sessionId.toString()
+            props["\$session_id"] = tempSessionId
+            if (config?.sessionReplay == true) {
+                // Session replay requires $window_id, so we set as the same as $session_id.
+                // the backend might fallback to $session_id if $window_id is not present next.
+                props["\$window_id"] = tempSessionId
             }
         }
 
@@ -697,25 +690,15 @@ public class PostHog private constructor(
     }
 
     override fun startSession() {
-        synchronized(sessionLock) {
-            if (sessionId == sessionIdNone) {
-                sessionId = TimeBasedEpochGenerator.generate()
-            }
-        }
+        PostHogSessionManager.startSession()
     }
 
     override fun endSession() {
-        synchronized(sessionLock) {
-            sessionId = sessionIdNone
-        }
+        PostHogSessionManager.endSession()
     }
 
     override fun isSessionActive(): Boolean {
-        var active: Boolean
-        synchronized(sessionLock) {
-            active = sessionId != sessionIdNone
-        }
-        return active
+        return PostHogSessionManager.isSessionActive()
     }
 
     override fun <T : PostHogConfig> getConfig(): T? {
